Add helper info on docker-start.sh. Allow user to change listen and port

+++ docker-start.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
+if [ -z "$1" ]; then
+  echo 'Usage:'
+  echo "  $0 ip:port"
+  echo 'Example: for only allow localhost (recommended)'
+  echo "  $0 127.0.0.1:8888"
+  echo 'Example: this will expose your port (use carefully)'
+  echo "  $0 0.0.0.0:8888"
+  exit 1
+fi
 WORKDIR=`pwd`
 
 #docker pull netivism/neticrm-ci:drone-php7
 docker rm -f neticrm-ci-php7
 docker run -d \
   --name neticrm-ci-php7 \
-  -p 8888:8080 \
+  -p $1:8080 \
   -v /etc/localtime:/etc/localtime:ro \
   -v $WORKDIR/container/init.sh:/init.sh \
   -v /mnt/neticrm-7/civicrm:/mnt/neticrm-7/civicrm \
